Flip utils/config dependency; standard rounding; trim comments
Per review:
- Flip the utils->config dependency so text_width can live in utils and
  be imported by config: move the test-only draw_and_compare and
  tikz_and_compare helpers (the only users of DRAWING_DEFAULT in utils)
  into test/drawing/drawing.py, so utils no longer imports config.
- Replace the dyadic-resolution rounding of text_width with standard
  rounding to three decimals (and return a plain float).
- Drop the verbose per-attribute comments in DRAWING_ATTRIBUTES and the
  redundant comments in drawing.py, keeping only the one explaining the
  0.5 wire margin.
