+++ test/drawing/tikz/bell-state.tikz
@@ -16,11 +16,11 @@
 \node [] (13) at (0.5, 0.75) {};
 \node [] (14) at (1.5, 1.25) {};
 \node [] (15) at (1.5, 0) {};
-\node [] (16) at (0.25, 4.25) {};
-\node [] (17) at (0.25, 4.75) {};
-\node [] (18) at (0.8125, 4.75) {};
-\node [] (19) at (0.8125, 4.25) {};
-\node [style=none, fill=white] (20) at (0.53125, 4.5) {sqrt(2)};
+\node [] (16) at (0.25000000000000006, 4.25) {};
+\node [] (17) at (0.25000000000000006, 4.75) {};
+\node [] (18) at (0.7890000000000001, 4.75) {};
+\node [] (19) at (0.7890000000000001, 4.25) {};
+\node [style=none, fill=white] (20) at (0.5195000000000001, 4.5) {sqrt(2)};
 \node [] (21) at (0.25, 3.25) {};
 \node [] (22) at (0.75, 3.25) {};
 \node [] (23) at (0.5, 3.75) {};
